Update wishes script.
I am new to Github, but I was able to clone the game as you said on Discord. As for the game, the first thing I did was to test out all the wishes list. There were some wordings and other stuff that seem a bit off to me, so I made sure to edit everything I can to make it better. If the writer or anyone disagrees with me, please tell me why.