Update wishes file #2
I was able to update parts of the wishes negative vibes for a bit. Although I can't really think of the negative vibes from the flying carpet and having a stallion. 
If my negative are boring and long. I'm sorry because I am a liberal arts major, I'm good with long, moral, and boring writing.  If you guys can make it more funny. Go ahead.